Build Monte Carlo (IVFN or TFN) FCMs

This function generates n fcm adjacency matrices whose edge weights are sampled from edge values (that may be Conventional, IVFNs, or TFNs) and stores them as a list of adjacency matrices.

Usage

build_monte_carlo_fcms_from_fuzzy_set_adj_matrices(
  fuzzy_set_adj_matrix_list = list(data.frame()),
  fuzzy_set_adj_matrix_list_class = c("conventional", "ivfn", "tfn"),
  N_samples = integer(),
  include_zeroes = FALSE,
  show_progress = TRUE
)

Arguments

fuzzy_set_adj_matrix_list

A list of n x n fuzzy adjacency matrices representing fcms

fuzzy_set_adj_matrix_list_class

"fgcm" or "fcm_w_tfn" - the class of elements in the fuzzy_set_adj_matrix_list

N_samples

The number of samples to draw from the corresponding distribution

include_zeroes

TRUE/FALSE Whether to incorporate zeroes as intentionally-defined edge weights or ignore them in aggregation

show_progress

TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply from the pbapply package as the underlying function.

Value

A list of empirical (Conventional) FCM adj. matrices generated via monte carlo methods

Details

If an edge is represented by IVFNs/TFNs, then those distributions are averaged together to create the aggregate distribution to sample from.

The show_progress and parallel inputs change the functions called, but do NOT change the output! These are allowed to be toggled on/off to increase user control at runtime.

See also

Examples

# For Interval-Valued Fuzzy Numbers (IVFNs)
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_1 <- make_adj_matrix_w_ivfns(lower_adj_matrix_1, upper_adj_matrix_1)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- make_adj_matrix_w_ivfns(lower_adj_matrix_2, upper_adj_matrix_2)
lower_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.2, 0)
)
upper_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_3 <- make_adj_matrix_w_ivfns(lower_adj_matrix_3, upper_adj_matrix_3)
lower_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.0, 0)
)
upper_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
adj_matrix_4 <- make_adj_matrix_w_ivfns(lower_adj_matrix_4, upper_adj_matrix_4)
fcms_w_ivfns <- list(adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4)

mc_ivfn_fcms <- build_monte_carlo_fcms_from_fuzzy_set_adj_matrices(
  fcms_w_ivfns, "ivfn", 1000, include_zeroes = FALSE
)
#> [1] Sampling from column vectors
#> Sampling from column vectors[1] Constructing monte carlo fcms from samples
#> Constructing monte carlo fcms from samples

# For Triangular Fuzzy Numbers (TFNs)
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
mode_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_1 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_1, mode_adj_matrix_1, upper_adj_matrix_1
)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
mode_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.7, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_2, mode_adj_matrix_2, upper_adj_matrix_2
)
lower_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.2, 0)
)
mode_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_3 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_3, mode_adj_matrix_3, upper_adj_matrix_3
)
lower_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.0, 0)
)
mode_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0, 0)
)
upper_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
adj_matrix_4 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_4, mode_adj_matrix_4, upper_adj_matrix_4
)

fcms_w_tfns <- list(
  adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4
)

mc_tfn_fcms <- build_monte_carlo_fcms_from_fuzzy_set_adj_matrices(
  fcms_w_tfns, "tfn", 1000, include_zeroes = FALSE
)
#> [1] Sampling from column vectors
#> Sampling from column vectors[1] Constructing monte carlo fcms from samples
#> Constructing monte carlo fcms from samples