Check Simulation Inputs

Confirm that all inputs will work with the simulation function and return appropriate error messages where necessary

Usage

check_simulation_inputs(
  adj_matrix = matrix(),
  initial_state_vector = c(),
  clamping_vector = c(),
  activation = c("kosko", "modified-kosko", "rescale"),
  squashing = c("sigmoid", "tanh"),
  lambda = 1,
  point_of_inference = c("peak", "final"),
  max_iter = 100,
  min_error = 1e-04
)

Arguments

adj_matrix

An n x n adjacency matrix that represents an FCM

initial_state_vector

A list state values at the start of an fcm simulation

clamping_vector

A list of values representing specific actions taken to control the behavior of an FCM. Specifically, non-zero values defined in this vector will remain constant throughout the entire simulation as if they were "clamped" at those values.

activation

The activation function to be applied. Must be one of the following: 'kosko', 'modified-kosko', or 'rescale'.

squashing

A squashing function to apply. Must be one of the following: 'tanh', or 'sigmoid'.

lambda

A numeric value that defines the steepness of the slope of the squashing function when tanh or sigmoid are applied

point_of_inference

The point along the simulation time-series to be identified as the inference. Must be one of the following: 'peak' or 'final'

max_iter

The maximum number of iterations to run if the minimum error value is not achieved

min_error

The lowest error (sum of the absolute value of the current state vector minus the previous state vector) at which no more iterations are necessary and the simulation will stop

Value

A formatted initial_state_vector and clamping_vector

Details

INTENDED FOR DEVELOPER USE ONLY

This checks that all inputs for a simulation function are of an appropriate format, and also fills in missing inputs for initial_state_vector, clamping_vector, and IDs when appropriate.

Examples

NULL
#> NULL