Convert IVFN or TFN Elements in Adj. Matrix to Distributions (i.e. sets)

Given a list of adjacency matrices which include either ivfns or tfns, convert those objects to their corresponding distributions representative of those values.

Usage

convert_fuzzy_set_elements_in_matrix_to_distributions(
  fuzzy_set_matrix = matrix(),
  object_class = c("ivfn", "tfn"),
  N_samples = integer()
)

Arguments

fuzzy_set_matrix

A matrix that contains fuzzy sets as elements

object_class

Values are represented either as ivfns or tfns. Options: 'ivfn' or 'tfn'

N_samples

The number of samples to draw from the corresponding distribution

Value

An adj. matrix of IVFNs or TFNs represented as lists (sets) of their representative distributions

Details

This function assists with subtracting the baseline from the scenario simulation when calling infer_fcm with IVFN-FCMs or TFN-FCMs.

Examples

lower_adj_matrix <- data.frame(
  C1 = c(0, 0, 0, 0, 0, 0),
  C2 = c(-0.85, 0, 0, 0.35, 0, 0),
  C3 = c(0, 0, 0, 0, 0, 0),
  C4 = c(-0.7, 0.6, -1, 0, -1, 0),
  C5 = c(0.1, 0, 0, -0.8, 0, 0),
  C6 = c(0, -0.95, 0, 0, -0.95, 0)
)
mode_adj_matrix <- data.frame(
  C1 = c(0, 0, 0, 0, 0, 0),
  C2 = c(-0.65, 0, 0, 0.80, 0, 0),
  C3 = c(0, 0, 0, 0, 0, 0),
  C4 = c(-0.6, 0.65, -0.85, 0, -0.95, 0),
  C5 = c(0.6, 0, 0, -0.2, 0, 0),
  C6 = c(0, -0.85, 0.1, 0, -0.75, 0)
)
upper_adj_matrix <- data.frame(
  C1 = c(0, 0, 0, 0, 0, 0),
  C2 = c(-0.1, 0, 0, 1, 0, 0),
  C3 = c(0, 0, 0, 0, 0, 0),
  C4 = c(-0.4, 0.7, -0.7, 0, -0.1, 0),
  C5 = c(0.7, 0, 0, -0.1, 0, 0),
  C6 = c(0, -0.2, 0.25, 0, -0.7, 0)
)
ex_tfn_adj_matrix <- make_adj_matrix_w_tfns(
  lower_adj_matrix, mode_adj_matrix, upper_adj_matrix
)

convert_fuzzy_set_elements_in_matrix_to_distributions(
  ex_tfn_adj_matrix, "tfn", 1000
)
#>    C1     C2     C3     C4     C5     C6    
#> C1 list,1 list,1 list,1 list,1 list,1 list,1
#> C2 list,1 list,1 list,1 list,1 list,1 list,1
#> C3 list,1 list,1 list,1 list,1 list,1 list,1
#> C4 list,1 list,1 list,1 list,1 list,1 list,1
#> C5 list,1 list,1 list,1 list,1 list,1 list,1
#> C6 list,1 list,1 list,1 list,1 list,1 list,1