View FCM Network

Display an FCM in the Viewer pane as an interactive visNetwork object. Use the shiny parameter to interact with the FCM visNetwork object and store its output in the global environment.

The shiny app provides the following functionalities: (1) interactive node placement, (2) the ability to toggle node labels and edge labels on and off, (3) the ability to change node size and shape, (4) the ability to change font size, (5) the ability to change the size of the arrowhead on directional edges, and (6) the ability to change the curvature of edges and the smoothing method used.

Usage

fcm_view(
  adj_matrix = data.frame(),
  fcm_visNetwork = list(),
  shiny = FALSE,
  ...
)

Arguments

adj_matrix

[list() or data.frame()]
A single adjacency matrix (n x n) representing FCMs. An adjacency matrix can have conventional edge weights, IVFN edge weights or TFN edge weights.

fcm_visNetwork

An fcm_view visNetwork object output. fcm_view accepts either an adj_matrix or fcm_visNetwork input but NOT both.

shiny

View visNetwork output in an interactive shiny app. Allows the user to manipulate and save node locations and other plot characteristics. Outputs are saved as fcm_visNetwork objects.

...

For advanced use. Set alert_on_open = FALSE to remove the alert pop-up that describes how to save visNetwork outputs.

Value

a visNetwork object that plots FCM networks in the Viewer pane; if shiny = TRUE, plots FCM networks in a shiny app

See also

Examples

adj_matrix <- sample_fcms$simple_fcms$conventional_fcms[[1]]
fcm_view(adj_matrix = adj_matrix, shiny = FALSE)

adj_matrix <- sample_fcms$simple_fcms$ivfn_fcms[[1]]
fcm_view(adj_matrix = adj_matrix, shiny = FALSE)

adj_matrix <- sample_fcms$simple_fcms$tfn_fcms[[1]]
fcm_view(adj_matrix = adj_matrix, shiny = FALSE)