fcmconfr_gui

fcmconfr_gui performs two tasks: (1) it launches a Shiny app that lets users interactively select parameters and (2) it outputs a corresponding call to fcmconfr that users can copy-and-paste to run in their own scripts.

No inputs are provided to fcmconfr_gui, but the local environment must have an individual FCM or a list object containing multiple FCMs in order for fcmconfr_gui to be used. Once fcmconfr_gui is running the user selects the appropriate FCMs to analyze from a dropdown menu titled Adj. Matrix or List of Adj. Matrices.

The GUI interface is organized into four tabs: Data, Agg. and Monte Carlo Options, Simulation Options, and Runtime Options, which are described further below.

Data: The Data tab is where the user selects the FCMs they want to analyze (drop-down menu) titled Adj. Matrix or List of Adj. Matrices). It is also where they manipulate values in the Initial State (Pulse) Vector or Clamping Vector to identify the dynamic simulations they wish to explore. A common form of simulation is a "clamped" simulation where every value in the Initial State (Pulse) Vector is set to 1 and only the node(s)-of-interest are set to 1 in the Clamping Vector. In this configuration, the Initial State (Pulse) Vector is used to determine the network's state at equilibrium and the Clamping Vector is used to determine the network's state in response to continuous activation of one or more nodes. The difference between the two is a measure of the impact of those particular nodes on the system.

Agg. and Monte Carlo Option: The Agg. and Monte Carlo Options tab is where the user specifies their preferences regarding FCM aggregation and Monte Carlo sampling. Aggregation options are only available when datasets contain multiple FCMs. Monte Carlo sampling is available for list objects of conventional, IVFN-, and TFN-FCMs as well as individual IVFN-/TFN-FCMs, but not individual conventional FCMs. Aggregation and Monte Carlo options can be toggled off to speed up the simulation process.

Aggregation options allow the user to specify whether multiplt FCMs should be aggregated into a single collective model so that dynamic simulations can be performed on the aggregate in addition to individual FCMs. The user can specify the aggregation method (mean or median) and whether zero-weighted edges should be included in the aggregation process.

Monte Carlo options allow the user to specify whether Monte Carlo sampling should be performed to estimate uncertainty in dynamic simulation outputs. If yes, the user must specify the number of FCMs that will be generated by Monte Carlo sampling (we recommend 1000 or more). Dynamic simulations will be performed on each of these FCMs, providing a range of inferences for each node, from which the median state and quantiles (25th, 75th) will be estimated. The user also has the option to perform nonparametric bootstrapping to estimate confidence bounds about the average inference for each node. Confidence limits are user specified (e.g., 95th percentile, 90th percentile, etc.).

Simulation Options: The Siimulation Options tab is where users specify the type of dynamic simulation to perform. This includes specifying (1) the activation function (Kosko, Modified-Kosko, or Rescale; default of Kosko), (2) the squashing function (sigmoid or hyperbolic tangent; default of sigmoid), (3) the squashing function's key parameter, lambda (default of 1), (4) the final simulation output (i.e., the peak estimate for each node or its final resting state; default of final) (5) the maximum number of iterations to perform per simulation (default of 100), and (6) the minimum acceptable error between iterations (default of \(1*10^{-5}\))

Runtime Options: The Runtime Options tab allows the user to specify whether they want to use parallel processing and have a progress bar displayed at runtime. These options do not impact simulation outputs. They do, however, influence how long fcmconfr takes to run and what the user sees at runtime.

A brief summary of each parameter within the fcmconfr_gui is provided in a glossary stored in a side tab within the GUI. The side tab can be opened by clicking the arrow symbol in the top-right-hand-corner of the GUI.

Usage

fcmconfr_gui()

Value

On exit, this function outputs a copy-and-paste sample script to call fcmconfr with the selected inputs.

Examples

NULL
#> NULL