Get the data types of adjacency matrices in a list

This function performs two actions:

1. This function identifies whether the input is a list of adjacency matrices or is an individual adj matrix (input_type)

2. This function identifies the 'class' of the input adj. matrices from the following options: 'conventional' 'ivfn' 'tfn' or 'unavailable'

- 'conventional' means that the adj. matrices contain only numeric objects

- 'ivfn' means that the adj. matrices contain only 'ivfn' objects (interval- valued fuzzy number) NOTE: also returns the matrix class type (i.e. data.frame, tibble, etc.)

- 'tfn' means that the adj. matrices contain only 'tfn' objects (triangular fuzzy number)

- 'unavailable' means that the adj. matrices contain non-numeric data that are not of types 'ivfn' or 'tfn'

Usage

get_adj_matrices_input_type(adj_matrix_list_input)

Arguments

adj_matrix_list_input

A list of adj matrices or an individual adj matrix

Value

a named list with two variables: adj_matrices_input_is_list: TRUE/FALSE Whether the input is a list of adj. matrices object_types_in_list: The 'class' of the input adj. matrices

See also

Examples

# Individual Conventional Adj. Matrix ----
adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
get_adj_matrices_input_type(adj_matrix_1)
#> $fcm_class
#> [1] "conventional"
#> 
#> $adj_matrices_input_is_list
#> [1] FALSE
#> 
#> $object_types_in_list
#> [1] "conventional" "data.frame"  
#> 

# List of Multiple Conventional FCM Adj. Matrices ----
adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.25, 0)
)
adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.75, 0)
)
adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.5, 0)
)
fcms <- list(adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4)
get_adj_matrices_input_type(fcms)
#> $fcm_class
#> [1] "conventional"
#> 
#> $adj_matrices_input_is_list
#> [1] TRUE
#> 
#> $object_types_in_list
#> [1] "conventional" "data.frame"  
#> 

# List of Multiple IVFN FCM Adj. Matrices ----
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_1 <- make_adj_matrix_w_ivfns(lower_adj_matrix_1, upper_adj_matrix_1)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- make_adj_matrix_w_ivfns(lower_adj_matrix_2, upper_adj_matrix_2)
lower_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.2, 0)
)
upper_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_3 <- make_adj_matrix_w_ivfns(lower_adj_matrix_3, upper_adj_matrix_3)
lower_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.0, 0)
)
upper_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
adj_matrix_4 <- make_adj_matrix_w_ivfns(lower_adj_matrix_4, upper_adj_matrix_4)
fcms_w_ivfns <- list(adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4)
get_adj_matrices_input_type(fcms_w_ivfns)
#> $fcm_class
#> [1] "ivfn"
#> 
#> $adj_matrices_input_is_list
#> [1] TRUE
#> 
#> $object_types_in_list
#> [1] "ivfn"
#> 

# List of Multiple TFN FCM Adj. Matrices ----
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
mode_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_1 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_1, mode_adj_matrix_1, upper_adj_matrix_1
)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
mode_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.7, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(1, 0)
)
adj_matrix_2 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_2, mode_adj_matrix_2, upper_adj_matrix_2
)
lower_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.2, 0)
)
mode_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
upper_adj_matrix_3 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.6, 0)
)
adj_matrix_3 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_3, mode_adj_matrix_3, upper_adj_matrix_3
)
lower_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.0, 0)
)
mode_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0, 0)
)
upper_adj_matrix_4 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.4, 0)
)
adj_matrix_4 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_4, mode_adj_matrix_4, upper_adj_matrix_4
)
fcms_w_tfns <- list(adj_matrix_1, adj_matrix_2, adj_matrix_3, adj_matrix_4)
get_adj_matrices_input_type(fcms_w_tfns)
#> $fcm_class
#> [1] "tfn"
#> 
#> $adj_matrices_input_is_list
#> [1] TRUE
#> 
#> $object_types_in_list
#> [1] "tfn"
#>