Simulate (Conventional) FCM

This simulates a conventional fcm based on its adjacency matrix.

Usage

simulate_conventional_fcm(
  adj_matrix = matrix(),
  initial_state_vector = c(),
  clamping_vector = c(),
  activation = "kosko",
  squashing = "tanh",
  lambda = 1,
  max_iter = 100,
  min_error = 1e-05
)

Arguments

adj_matrix

An n x n adjacency matrix that represents an FCM

initial_state_vector

A list state values at the start of an fcm simulation

clamping_vector

A list of values representing specific actions taken to control the behavior of an FCM. Specifically, non-zero values defined in this vector will remain constant throughout the entire simulation as if they were "clamped" at those values.

activation

The activation function to be applied. Must be one of the following: 'kosko', 'modified-kosko', or 'rescale'.

squashing

A squashing function to apply. Must be one of the following: 'bivalent', 'saturation', 'trivalent', 'tanh', or 'sigmoid'.

lambda

A numeric value that defines the steepness of the slope of the squashing function when tanh or sigmoid are applied

max_iter

The maximum number of iterations to run if the minimum error value is not achieved

min_error

The lowest error (sum of the absolute value of the current state vector minus the previous state vector) at which no more iterations are necessary and the simulation will stop

Value

(Conventional) FCM simulation results

Details

This function performs two fcm simulations and compares the output between the two. The first simulation considers the baseline activity where no nodes are "clamped" and the system behaves without any outside inputs. The second simulation considers a scenario where one or multiple nodes are "clamped" so that the system is reactive to additional inputs. The function returns the difference in simulation results between the scenario and baseline activity to understand how system manipulations compare to structural expectations of the system.

This function produces the same output as mental modeler for the following inputs: - initial_state_vector = c(1, 1, ..., 1) - activation = "kosko" - squashing = either "sigmoid" or "tanh" - lambda = 1

Examples

adj_matrix <- data.frame(
  C1 = c(0, 0, 0, 0, 0, 0),
  C2 = c(-0.85, 0, 0, 0.35, 0, 0),
  C3 = c(0, 0, 0, 0, 0, 0),
  C4 = c(-0.7, 0.6, -1, 0, -1, 0),
  C5 = c(0.1, 0, 0, -0.8, 0, 0),
  C6 = c(0, -0.95, 0, 0, -0.95, 0)
)

simulate_fcm(adj_matrix,
             initial_state_vector = c(1, 1, 1, 1, 1, 1),
             clamping_vector = c(1, 0, 0, 0, 0, 0),
             activation = "kosko",
             squashing = "sigmoid",
             lambda = 1)
#> $state_vectors
#>    iter C1        C2  C3        C4        C5        C6
#> 1     0  1 1.0000000 1.0 1.0000000 1.0000000 1.0000000
#> 2     1  1 0.3775407 0.5 0.1090968 0.3318122 0.1301085
#> 3     2  1 0.3075036 0.5 0.2132764 0.5031806 0.3376274
#> 4     3  1 0.3153222 0.5 0.1796562 0.4823520 0.3164467
#> 5     4  1 0.3127873 0.5 0.1834481 0.4890705 0.3191262
#> 6     5  1 0.3130727 0.5 0.1822171 0.4883125 0.3182632
#> 7     6  1 0.3129800 0.5 0.1823556 0.4885586 0.3183606
#> 8     7  1 0.3129904 0.5 0.1823106 0.4885309 0.3183290
#> 9     8  1 0.3129871 0.5 0.1823157 0.4885399 0.3183326
#> 10    9  1 0.3129874 0.5 0.1823140 0.4885389 0.3183314
#> 
#> $errors
#>    iter C1           C2  C3           C4           C5           C6
#> 1     0  0 0.000000e+00 0.0 0.000000e+00 0.000000e+00 0.000000e+00
#> 2     1  0 6.224593e-01 0.5 8.909032e-01 6.681878e-01 8.698915e-01
#> 3     2  0 7.003704e-02 0.0 1.041796e-01 1.713684e-01 2.075189e-01
#> 4     3  0 7.818602e-03 0.0 3.362022e-02 2.082854e-02 2.118076e-02
#> 5     4  0 2.534905e-03 0.0 3.791884e-03 6.718451e-03 2.679494e-03
#> 6     5  0 2.853456e-04 0.0 1.231001e-03 7.579887e-04 8.629478e-04
#> 7     6  0 9.265046e-05 0.0 1.385041e-04 2.460686e-04 9.743062e-05
#> 8     7  0 1.042367e-05 0.0 4.497364e-05 2.768629e-05 3.162735e-05
#> 9     8  0 3.384688e-06 0.0 5.059681e-06 8.989999e-06 3.558627e-06
#> 10    9  0 3.807874e-07 0.0 1.642940e-06 1.011404e-06 1.155517e-06
#> 
#> $params
#> $params$adj_matrix
#>   C1    C2 C3   C4   C5    C6
#> 1  0 -0.85  0 -0.7  0.1  0.00
#> 2  0  0.00  0  0.6  0.0 -0.95
#> 3  0  0.00  0 -1.0  0.0  0.00
#> 4  0  0.35  0  0.0 -0.8  0.00
#> 5  0  0.00  0 -1.0  0.0 -0.95
#> 6  0  0.00  0  0.0  0.0  0.00
#> 
#> $params$initial_state_vector
#> [1] 1 1 1 1 1 1
#> 
#> $params$activation
#> [1] "kosko"
#> 
#> $params$squashing
#> [1] "sigmoid"
#> 
#> $params$lambda
#> [1] 1
#> 
#> $params$max_iter
#> [1] 100
#> 
#> $params$min_error
#> [1] 1e-05
#> 
#> $params$concepts
#> [1] "C1" "C2" "C3" "C4" "C5" "C6"
#> 
#> 
#> attr(,"class")
#> [1] "fcm_simulation"

# To simulate an fcm without clamping (i.e. don't calculate the difference
# from steady-state), change the values in the initial_state_vector while
# leaving all values in the clamping_vector at 0
simulate_fcm(adj_matrix,
             initial_state_vector = c(1, 0, 0, 0, 0, 0),
             clamping_vector = c(0, 0, 0, 0, 0, 0),
             activation = "kosko",
             squashing = "sigmoid",
             lambda = 1)
#> $state_vectors
#>    iter  C1        C2  C3        C4        C5        C6
#> 1     0 1.0 0.0000000 0.0 0.0000000 0.0000000 0.0000000
#> 2     1 0.5 0.2994329 0.5 0.3318122 0.5249792 0.5000000
#> 3     2 0.5 0.4233916 0.5 0.2323089 0.4463449 0.3136325
#> 4     3 0.5 0.4149130 0.5 0.2607045 0.4660904 0.3044386
#> 5     4 0.5 0.4173276 0.5 0.2559469 0.4604419 0.3021769
#> 6     5 0.5 0.4169228 0.5 0.2573008 0.4613877 0.3028251
#> 7     6 0.5 0.4170380 0.5 0.2570737 0.4611185 0.3027166
#> 8     7 0.5 0.4170187 0.5 0.2571383 0.4611636 0.3027475
#> 9     8 0.5 0.4170242 0.5 0.2571275 0.4611508 0.3027423
#> 10    9 0.5 0.4170232 0.5 0.2571306 0.4611530 0.3027438
#> 
#> $errors
#>    iter  C1           C2  C3           C4           C5           C6
#> 1     0 0.0 0.000000e+00 0.0 0.000000e+00 0.000000e+00 0.000000e+00
#> 2     1 0.5 2.994329e-01 0.5 3.318122e-01 5.249792e-01 5.000000e-01
#> 3     2 0.0 1.239588e-01 0.0 9.950333e-02 7.863424e-02 1.863675e-01
#> 4     3 0.0 8.478674e-03 0.0 2.839563e-02 1.974541e-02 9.193839e-03
#> 5     4 0.0 2.414687e-03 0.0 4.757664e-03 5.648411e-03 2.261750e-03
#> 6     5 0.0 4.048587e-04 0.0 1.353921e-03 9.457181e-04 6.481817e-04
#> 7     6 0.0 1.152020e-04 0.0 2.270788e-04 2.691580e-04 1.084670e-04
#> 8     7 0.0 1.932225e-05 0.0 6.461218e-05 4.514145e-05 3.087296e-05
#> 9     8 0.0 5.497857e-06 0.0 1.083723e-05 1.284445e-05 5.177677e-06
#> 10    9 0.0 9.221424e-07 0.0 3.083565e-06 2.154362e-06 1.473250e-06
#> 
#> $params
#> $params$adj_matrix
#>   C1    C2 C3   C4   C5    C6
#> 1  0 -0.85  0 -0.7  0.1  0.00
#> 2  0  0.00  0  0.6  0.0 -0.95
#> 3  0  0.00  0 -1.0  0.0  0.00
#> 4  0  0.35  0  0.0 -0.8  0.00
#> 5  0  0.00  0 -1.0  0.0 -0.95
#> 6  0  0.00  0  0.0  0.0  0.00
#> 
#> $params$initial_state_vector
#> [1] 1 0 0 0 0 0
#> 
#> $params$activation
#> [1] "kosko"
#> 
#> $params$squashing
#> [1] "sigmoid"
#> 
#> $params$lambda
#> [1] 1
#> 
#> $params$max_iter
#> [1] 100
#> 
#> $params$min_error
#> [1] 1e-05
#> 
#> $params$concepts
#> [1] "C1" "C2" "C3" "C4" "C5" "C6"
#> 
#> 
#> attr(,"class")
#> [1] "fcm_simulation"