Display FCM in Viewer

Display an FCM in the Viewer pane as an interactive visNetwork object. Use the with_shiny parameter to interact with the FCM visNetwork object and store its output in the global environment via the view_fcm_visNetwork variable.

Usage

fcm_view(fcm_adj_matrix = matrix(), with_shiny = FALSE)

Arguments

fcm_adj_matrix

An adjacency matrix representing an FCM

with_shiny

View visNetwork output in a shiny app instead of the viewer pane. This adds the functionality to store node locations after moving them.

Value

A display of the input FCM in the Viewer pane; if with_shiny = TRUE, the visNetwork object saved into the global environment as fcm_view_visNetwork

See also

Examples

adj_matrix <- sample_fcms$simple_fcms$conventional_fcms[[1]]
fcm_view(fcm_adj_matrix = adj_matrix)

adj_matrix <- sample_fcms$simple_fcms$ivfn_fcms[[1]]
fcm_view(fcm_adj_matrix = adj_matrix)

adj_matrix <- sample_fcms$simple_fcms$tfn_fcms[[1]]
fcm_view(fcm_adj_matrix = adj_matrix)