fcmconfr

This is the primary function of the fcmconfr package. This function performs up to three different analyses of a set of (or an individual) input FCM(s). Use fcmconfr_gui to help select fcmconfr parameters using an interactive GUI.

FCM Simulation: Simulate the (raw/unmodified) input FCMs to estimate their inferences. Inferences are calculated by comparing the simulation output of a particular scenario with the simulation output of the baseline (i.e. the natural behavior of the network without external perturbation), (Ozesmi & Ozesmi, 2003) and help distinguish decision-making impacts with the structural or expected, steady-state of the system.

Aggregate Analysis: Generate an aggregate adj. matrix from a list of adj. matrices. FCM aggregation works by calculating the mean/median edge weight for all edges across the input adj. matrices (i.e. the mean/median of the edge weight connecting A -> B across all maps, the mean/median of the edge weight connecting B -> C across all maps, and so on) (Aminpour et al., 2020). The user may dictate whether to incorporate 0-valued edge weights in the mean/median calculations. See aggregate_fcms for more information on aggregating FCMs outside of the primary fcmconfr function.

Monte Carlo Analysis: Generate empirical adj. matrices whose edge weights are drawn from edge weights from the input FCMs using monte carlo sampling methods.These empirical FCMs are then simulated in bulk using the FCM Simulation method previously described. This outputs a dataframe of inferences across all empirical FCMs that represent the possibility space of inferences representative of the collective. This function also uses bootstrapping methods to estimate confidence intervals about the inferences generated from the empirical FCMs (may be toggled off if wanting to reduce runtime).

This function accepts three different types of FCMs which differ in how they represent edge weights. Note: All input FCMs must be the same type (i.e. a list of input FCMs must all be of one type or another, but cannot contain multiple types of FCMs in the same input set).

Conventional FCMs: These represent edge weights as fuzzy numbers and represent FCMs in thir traditional form (Stylios, 1997)

Interval-Valued Fuzzy Number (IVFN) FCMs: An extension of Conventional FCMs, these represent edge weights as interval-valued fuzzy numbers (IVFNs) which describe a range of values and are defined by a lower and upper bound (e.g. [0.2, 0.8] represents the set of values between 0.2 and 0.8 which may be represented as a uniform distribution via runif(N, min = 0.2, max = 0.8)) (Moore & Lodwick, 2003)

Triangular Fuzzy Number (TFN) FCMs: An extension of Conventional FCMs, these represent edge weights as triangular fuzzy numbers (TFNs) which describe a bounded range of values with a greater density of values at some point in between, such that the probability density function of the distribution appears triangular (Moore & Lodwick, 2003). These are defined by their lower and upper bounds as well as a mode value, and can be called within this package via the rtriangular_dist function or with other packages such as EnvStats::rtri().

Usage

fcmconfr(
  adj_matrices = list(matrix()),
  agg_function = c("mean", "median"),
  num_mc_fcms = 1000,
  initial_state_vector = c(),
  clamping_vector = c(),
  activation = c("kosko", "modified-kosko", "rescale"),
  squashing = c("sigmoid", "tanh"),
  lambda = 1,
  point_of_inference = c("peak", "final"),
  max_iter = 100,
  min_error = 1e-05,
  ci_centering_function = c("mean", "median"),
  confidence_interval = 0.95,
  num_ci_bootstraps = 5000,
  show_progress = TRUE,
  parallel = FALSE,
  n_cores = integer(),
  run_agg_calcs = TRUE,
  run_mc_calcs = TRUE,
  run_ci_calcs = TRUE,
  include_zeroes_in_sampling = FALSE,
  include_sims_in_output = TRUE
)

Arguments

adj_matrices

A list of adjacency matrices (n x n) representing FCMs. This can also be an individual adjacency matrix.Adj. Matrices can be conventional FCMs, FCMs with edge weights as Interval Value Fuzzy Numbers (IVFNs) or FCMs with edge weights as Triangular Fuzzy Numbers (TFNs)

agg_function

Aggregate the adj. matrices into a single FCM by taking either the mean or median of the edge weights for edges included in multiple maps

num_mc_fcms

The number of FCMs to generate via monte carlo sampling from the input adj. matrices

initial_state_vector

A list state values at the start of an fcm simulation

clamping_vector

A list of values representing specific actions taken to control the behavior of an FCM. Specifically, non-zero values defined in this vector will remain constant throughout the entire simulation as if they were "clamped" at those values.

activation

The activation function to be applied. Must be one of the following: 'kosko', 'modified-kosko', or 'rescale'.

squashing

A squashing function to apply. Must be one of the following: 'tanh', or 'sigmoid'.

lambda

A numeric value that defines the steepness of the slope of the squashing function when tanh or sigmoid are applied

point_of_inference

The point along the simulation time-series to be identified as the inference. Must be one of the following: 'peak' or 'final'

max_iter

The maximum number of iterations to run if the minimum error value is not achieved

min_error

The lowest error (sum of the absolute value of the current state vector minus the previous state vector) at which no more iterations are necessary and the simulation will stop

ci_centering_function

Estimate confidence intervals about the "mean" or "median" of inferences from the monte carlo simulations

confidence_interval

The confidence interval to estimate for the inferences of each concept across all monte carlo FCMs (via bootstrap)

num_ci_bootstraps

The number of bootstraps to perform in estimating the confidence interval for the inferences of each concept across all monte carlo FCMs

show_progress

TRUE/FALSE Show progress bar when creating fmcm. Uses pbmapply from the pbapply package as the underlying function.

parallel

TRUE/FALSE Whether to utilize parallel processing

n_cores

Number of cores to use in parallel processing. If no input given, will use all available cores in the machine.

run_agg_calcs

TRUE/FALSE Run the code to generate and simulate an aggregate FCM generated from the input adj_matrices

run_mc_calcs

TRUE/FALSE Run the code to generate and simulate monte carlo-generated FCM sampled from the input adj_matrices

run_ci_calcs

TRUE/FALSE Run the code to estimate the 95 percent CI bounds about the means of the inferences of the monte carlo adj matrices

include_zeroes_in_sampling

TRUE/FALSE Whether to incorporate zeroes as intentionally-defined edge weights or ignore them when aggregating adj. matrices and sampling for monte carlo FCMs

include_sims_in_output

TRUE/FALSE Whether to include simulations of monte carlo FCMs. Switch to FALSE if concerned about the size of the output of fcmconfr (simulations are necessary and will run regardless)

Value

A list of outputs generated from the individual_fcms simulations, aggregate_fcm analysis, and monte_carlo_fcms analysis. Bootstrap estimates of inferences from monte carlo analysis are included, as well as function inputs.

Details

Call fcmconfr_gui. for assistance with inputs.

References

ozesmiParticipatoryApproachEcosystem2003fcmconfr

aminpourWisdomStakeholderCrowds2020fcmconfr

styliosIntroducingTheoryFuzzy1997fcmconfr

mooreIntervalAnalysisFuzzy2003fcmconfr

Examples

# Conventional FCMs
test_adj_matrix_1 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(1, 0, 0, 1),
  "C" = c(0, 1, 0, 0),
  "D" = c(0, 0, 1, 0)
)
test_adj_matrix_2 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.25, 0, 0, 0.25),
  "C" = c(0, 0.25, 0, 0),
  "D" = c(0, 0, 0.25, 0)
)
test_adj_matrix_3 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.75, 0, 0, 0.75),
  "C" = c(0, 0.75, 0, 0),
  "D" = c(0, 0, 0.75, 0)
)
test_adj_matrix_4 <- data.frame(
  "A" = c(0, 0, 0, 0),
  "B" = c(0.5, 0, 0, 0.5),
  "C" = c(0, 0.5, 0, 0),
  "D" = c(0, 0, 0.5, 0)
)
ex_conventional_fcms <- list(
  test_adj_matrix_1, test_adj_matrix_2, test_adj_matrix_3, test_adj_matrix_4
)
ex_conventional_fcmconfr <- fcmconfr(
  adj_matrices = ex_conventional_fcms,
  # Aggregation and Monte Carlo Sampling
  agg_function = 'mean',
  num_mc_fcms = 100,
  # Simulation
  initial_state_vector = c(1, 1, 1, 1),
  clamping_vector = c(0, 1, 0, 0),
  activation = 'kosko',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 100,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  ci_centering_function = "median",
  confidence_interval = 0.95,
  num_ci_bootstraps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  run_agg_calcs = TRUE,
  run_mc_calcs = TRUE,
  run_ci_calcs = TRUE,
  include_zeroes_in_sampling = FALSE,
  include_sims_in_output = TRUE
)
#> [1] Simulating Input FCMs
#> [1] Initializing cluster
#> [1] Running Simulations in Parallel
#> [1] Sampling from column vectors
#> Sampling from column vectors[1] Constructing monte carlo fcms from samples
#> Constructing monte carlo fcms from samples[1] Initializing cluster
#> [1] Running Simulations in Parallel
#> [1] Performing bootstrap simulations
#> [1] Initializing cluster
#> [1] Sampling means
#>   |                                                          |                                                  |   1%  |                                                          |=                                                 |   1%  |                                                          |=                                                 |   2%  |                                                          |=                                                 |   3%  |                                                          |==                                                |   3%  |                                                          |==                                                |   4%  |                                                          |==                                                |   5%  |                                                          |===                                               |   5%  |                                                          |===                                               |   6%  |                                                          |===                                               |   7%  |                                                          |====                                              |   7%  |                                                          |====                                              |   8%  |                                                          |====                                              |   9%  |                                                          |=====                                             |   9%  |                                                          |=====                                             |  10%  |                                                          |=====                                             |  11%  |                                                          |======                                            |  11%  |                                                          |======                                            |  12%  |                                                          |======                                            |  13%  |                                                          |=======                                           |  13%  |                                                          |=======                                           |  14%  |                                                          |=======                                           |  15%  |                                                          |========                                          |  15%  |                                                          |========                                          |  16%  |                                                          |========                                          |  17%  |                                                          |=========                                         |  17%  |                                                          |=========                                         |  18%  |                                                          |=========                                         |  19%  |                                                          |==========                                        |  19%  |                                                          |==========                                        |  20%  |                                                          |==========                                        |  21%  |                                                          |===========                                       |  21%  |                                                          |===========                                       |  22%  |                                                          |===========                                       |  23%  |                                                          |============                                      |  23%  |                                                          |============                                      |  24%  |                                                          |============                                      |  25%  |                                                          |=============                                     |  25%  |                                                          |=============                                     |  26%  |                                                          |=============                                     |  27%  |                                                          |==============                                    |  27%  |                                                          |==============                                    |  28%  |                                                          |==============                                    |  29%  |                                                          |===============                                   |  29%  |                                                          |===============                                   |  30%  |                                                          |===============                                   |  31%  |                                                          |================                                  |  31%  |                                                          |================                                  |  32%  |                                                          |================                                  |  33%  |                                                          |=================                                 |  33%  |                                                          |=================                                 |  34%  |                                                          |=================                                 |  35%  |                                                          |==================                                |  35%  |                                                          |==================                                |  36%  |                                                          |==================                                |  37%  |                                                          |===================                               |  37%  |                                                          |===================                               |  38%  |                                                          |===================                               |  39%  |                                                          |====================                              |  39%  |                                                          |====================                              |  40%  |                                                          |====================                              |  41%  |                                                          |=====================                             |  41%  |                                                          |=====================                             |  42%  |                                                          |=====================                             |  43%  |                                                          |======================                            |  43%  |                                                          |======================                            |  44%  |                                                          |======================                            |  45%  |                                                          |=======================                           |  45%  |                                                          |=======================                           |  46%  |                                                          |=======================                           |  47%  |                                                          |========================                          |  47%  |                                                          |========================                          |  48%  |                                                          |========================                          |  49%  |                                                          |=========================                         |  49%  |                                                          |=========================                         |  50%  |                                                          |=========================                         |  51%  |                                                          |==========================                        |  51%  |                                                          |==========================                        |  52%  |                                                          |==========================                        |  53%  |                                                          |===========================                       |  53%  |                                                          |===========================                       |  54%  |                                                          |===========================                       |  55%  |                                                          |============================                      |  55%  |                                                          |============================                      |  56%  |                                                          |============================                      |  57%  |                                                          |=============================                     |  57%  |                                                          |=============================                     |  58%  |                                                          |=============================                     |  59%  |                                                          |==============================                    |  59%  |                                                          |==============================                    |  60%  |                                                          |==============================                    |  61%  |                                                          |===============================                   |  61%  |                                                          |===============================                   |  62%  |                                                          |===============================                   |  63%  |                                                          |================================                  |  63%  |                                                          |================================                  |  64%  |                                                          |================================                  |  65%  |                                                          |=================================                 |  65%  |                                                          |=================================                 |  66%  |                                                          |=================================                 |  67%  |                                                          |==================================                |  67%  |                                                          |==================================                |  68%  |                                                          |==================================                |  69%  |                                                          |===================================               |  69%  |                                                          |===================================               |  70%  |                                                          |===================================               |  71%  |                                                          |====================================              |  71%  |                                                          |====================================              |  72%  |                                                          |====================================              |  73%  |                                                          |=====================================             |  73%  |                                                          |=====================================             |  74%  |                                                          |=====================================             |  75%  |                                                          |======================================            |  75%  |                                                          |======================================            |  76%  |                                                          |======================================            |  77%  |                                                          |=======================================           |  77%  |                                                          |=======================================           |  78%  |                                                          |=======================================           |  79%  |                                                          |========================================          |  79%  |                                                          |========================================          |  80%  |                                                          |========================================          |  81%  |                                                          |=========================================         |  81%  |                                                          |=========================================         |  82%  |                                                          |=========================================         |  83%  |                                                          |==========================================        |  83%  |                                                          |==========================================        |  84%  |                                                          |==========================================        |  85%  |                                                          |===========================================       |  85%  |                                                          |===========================================       |  86%  |                                                          |===========================================       |  87%  |                                                          |============================================      |  87%  |                                                          |============================================      |  88%  |                                                          |============================================      |  89%  |                                                          |=============================================     |  89%  |                                                          |=============================================     |  90%  |                                                          |=============================================     |  91%  |                                                          |==============================================    |  91%  |                                                          |==============================================    |  92%  |                                                          |==============================================    |  93%  |                                                          |===============================================   |  93%  |                                                          |===============================================   |  94%  |                                                          |===============================================   |  95%  |                                                          |================================================  |  95%  |                                                          |================================================  |  96%  |                                                          |================================================  |  97%  |                                                          |================================================= |  97%  |                                                          |================================================= |  98%  |                                                          |================================================= |  99%  |                                                          |==================================================|  99%  |                                                          |==================================================| 100%
#> [1] Done

# IVFN FCM fcmconfr
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.25, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.75, 0)
)
adj_matrix_w_ivfns_1 <- make_adj_matrix_w_ivfns(
  lower_adj_matrix_1, upper_adj_matrix_1
)
lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.15, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.85, 0)
)
adj_matrix_w_ivfns_2 <- make_adj_matrix_w_ivfns(
  lower_adj_matrix_2, upper_adj_matrix_2
)
ex_ivfn_fcms <- list(adj_matrix_w_ivfns_1, adj_matrix_w_ivfns_2)
ex_ivfn_fcmconfr <- fcmconfr(
  adj_matrices = ex_ivfn_fcms,
  # Aggregation and Monte Carlo Sampling
  agg_function = 'mean',
  num_mc_fcms = 100,
  # Simulation
  initial_state_vector = c(1, 1),
  clamping_vector = c(0, 1),
  activation = 'kosko',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 100,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  ci_centering_function = "mean",
  confidence_interval = 0.95,
  num_ci_bootstraps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = TRUE,
  n_cores = 2,
  # Additional Options
  run_agg_calcs = TRUE,
  run_mc_calcs = TRUE,
  run_ci_calcs = TRUE,
  include_zeroes_in_sampling = FALSE,
  include_sims_in_output = TRUE
)
#> [1] Simulating Input FCMs
#> [1] Initializing cluster
#> [1] Running Simulations in Parallel
#> [1] Sampling from column vectors
#> Sampling from column vectors[1] Constructing monte carlo fcms from samples
#> Constructing monte carlo fcms from samples[1] Initializing cluster
#> [1] Running Simulations in Parallel
#> [1] Performing bootstrap simulations
#> [1] Initializing cluster
#> [1] Sampling means
#>   |                                                          |                                                  |   1%  |                                                          |=                                                 |   1%  |                                                          |=                                                 |   2%  |                                                          |=                                                 |   3%  |                                                          |==                                                |   3%  |                                                          |==                                                |   4%  |                                                          |==                                                |   5%  |                                                          |===                                               |   5%  |                                                          |===                                               |   6%  |                                                          |===                                               |   7%  |                                                          |====                                              |   7%  |                                                          |====                                              |   8%  |                                                          |====                                              |   9%  |                                                          |=====                                             |   9%  |                                                          |=====                                             |  10%  |                                                          |=====                                             |  11%  |                                                          |======                                            |  11%  |                                                          |======                                            |  12%  |                                                          |======                                            |  13%  |                                                          |=======                                           |  13%  |                                                          |=======                                           |  14%  |                                                          |=======                                           |  15%  |                                                          |========                                          |  15%  |                                                          |========                                          |  16%  |                                                          |========                                          |  17%  |                                                          |=========                                         |  17%  |                                                          |=========                                         |  18%  |                                                          |=========                                         |  19%  |                                                          |==========                                        |  19%  |                                                          |==========                                        |  20%  |                                                          |==========                                        |  21%  |                                                          |===========                                       |  21%  |                                                          |===========                                       |  22%  |                                                          |===========                                       |  23%  |                                                          |============                                      |  23%  |                                                          |============                                      |  24%  |                                                          |============                                      |  25%  |                                                          |=============                                     |  25%  |                                                          |=============                                     |  26%  |                                                          |=============                                     |  27%  |                                                          |==============                                    |  27%  |                                                          |==============                                    |  28%  |                                                          |==============                                    |  29%  |                                                          |===============                                   |  29%  |                                                          |===============                                   |  30%  |                                                          |===============                                   |  31%  |                                                          |================                                  |  31%  |                                                          |================                                  |  32%  |                                                          |================                                  |  33%  |                                                          |=================                                 |  33%  |                                                          |=================                                 |  34%  |                                                          |=================                                 |  35%  |                                                          |==================                                |  35%  |                                                          |==================                                |  36%  |                                                          |==================                                |  37%  |                                                          |===================                               |  37%  |                                                          |===================                               |  38%  |                                                          |===================                               |  39%  |                                                          |====================                              |  39%  |                                                          |====================                              |  40%  |                                                          |====================                              |  41%  |                                                          |=====================                             |  41%  |                                                          |=====================                             |  42%  |                                                          |=====================                             |  43%  |                                                          |======================                            |  43%  |                                                          |======================                            |  44%  |                                                          |======================                            |  45%  |                                                          |=======================                           |  45%  |                                                          |=======================                           |  46%  |                                                          |=======================                           |  47%  |                                                          |========================                          |  47%  |                                                          |========================                          |  48%  |                                                          |========================                          |  49%  |                                                          |=========================                         |  49%  |                                                          |=========================                         |  50%  |                                                          |=========================                         |  51%  |                                                          |==========================                        |  51%  |                                                          |==========================                        |  52%  |                                                          |==========================                        |  53%  |                                                          |===========================                       |  53%  |                                                          |===========================                       |  54%  |                                                          |===========================                       |  55%  |                                                          |============================                      |  55%  |                                                          |============================                      |  56%  |                                                          |============================                      |  57%  |                                                          |=============================                     |  57%  |                                                          |=============================                     |  58%  |                                                          |=============================                     |  59%  |                                                          |==============================                    |  59%  |                                                          |==============================                    |  60%  |                                                          |==============================                    |  61%  |                                                          |===============================                   |  61%  |                                                          |===============================                   |  62%  |                                                          |===============================                   |  63%  |                                                          |================================                  |  63%  |                                                          |================================                  |  64%  |                                                          |================================                  |  65%  |                                                          |=================================                 |  65%  |                                                          |=================================                 |  66%  |                                                          |=================================                 |  67%  |                                                          |==================================                |  67%  |                                                          |==================================                |  68%  |                                                          |==================================                |  69%  |                                                          |===================================               |  69%  |                                                          |===================================               |  70%  |                                                          |===================================               |  71%  |                                                          |====================================              |  71%  |                                                          |====================================              |  72%  |                                                          |====================================              |  73%  |                                                          |=====================================             |  73%  |                                                          |=====================================             |  74%  |                                                          |=====================================             |  75%  |                                                          |======================================            |  75%  |                                                          |======================================            |  76%  |                                                          |======================================            |  77%  |                                                          |=======================================           |  77%  |                                                          |=======================================           |  78%  |                                                          |=======================================           |  79%  |                                                          |========================================          |  79%  |                                                          |========================================          |  80%  |                                                          |========================================          |  81%  |                                                          |=========================================         |  81%  |                                                          |=========================================         |  82%  |                                                          |=========================================         |  83%  |                                                          |==========================================        |  83%  |                                                          |==========================================        |  84%  |                                                          |==========================================        |  85%  |                                                          |===========================================       |  85%  |                                                          |===========================================       |  86%  |                                                          |===========================================       |  87%  |                                                          |============================================      |  87%  |                                                          |============================================      |  88%  |                                                          |============================================      |  89%  |                                                          |=============================================     |  89%  |                                                          |=============================================     |  90%  |                                                          |=============================================     |  91%  |                                                          |==============================================    |  91%  |                                                          |==============================================    |  92%  |                                                          |==============================================    |  93%  |                                                          |===============================================   |  93%  |                                                          |===============================================   |  94%  |                                                          |===============================================   |  95%  |                                                          |================================================  |  95%  |                                                          |================================================  |  96%  |                                                          |================================================  |  97%  |                                                          |================================================= |  97%  |                                                          |================================================= |  98%  |                                                          |================================================= |  99%  |                                                          |==================================================|  99%  |                                                          |==================================================| 100%
#> [1] Done

# TFN FCM fcmconfr
lower_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.25, 0)
)
mode_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.5, 0)
)
upper_adj_matrix_1 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.75, 0)
)
adj_matrix_w_tfns_1 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_1, mode_adj_matrix_1, upper_adj_matrix_1
)

lower_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.15, 0)
)
mode_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.65, 0)
)
upper_adj_matrix_2 <- data.frame(
  "A" = c(0, 0),
  "B" = c(0.85, 0)
)
adj_matrix_w_tfns_2 <- make_adj_matrix_w_tfns(
  lower_adj_matrix_2, mode_adj_matrix_2, upper_adj_matrix_2
)

ex_tfn_fcms <- list(adj_matrix_w_tfns_1, adj_matrix_w_tfns_2)
ex_tfn_fcmconfr <- fcmconfr(
  adj_matrices = ex_tfn_fcms,
  # Aggregation and Monte Carlo Sampling
  agg_function = 'mean',
  num_mc_fcms = 100,
  # Simulation
  initial_state_vector = c(1, 1),
  clamping_vector = c(0, 1),
  activation = 'kosko',
  squashing = 'sigmoid',
  lambda = 1,
  point_of_inference = "final",
  max_iter = 100,
  min_error = 1e-05,
  # Inference Estimation (bootstrap)
  ci_centering_function = "mean",
  confidence_interval = 0.95,
  num_ci_bootstraps = 1000,
  # Runtime Options
  show_progress = TRUE,
  parallel = FALSE,
  # Additional Options
  run_agg_calcs = TRUE,
  run_mc_calcs = TRUE,
  run_ci_calcs = TRUE,
  include_zeroes_in_sampling = FALSE,
  include_sims_in_output = TRUE
)
#> [1] Simulating Input FCMs
#> 
#> [1] Running Simulations
#> [1] Sampling from column vectors
#> Sampling from column vectors[1] Constructing monte carlo fcms from samples
#> Constructing monte carlo fcms from samples
#> [1] Running Simulations
#> [1] Done